Optimized performance and fixed more bugs

+++ src/main/java/me/koenn/kingdomwars/stats/Level.java
@@ -20,8 +20,8 @@ public class Level {
     }
 
     public Level(JSONObject json) {
-        this.level = (int) json.get("level");
-        this.exp = (int) json.get("exp");
+        this.level = Math.toIntExact((long) json.get("level"));
+        this.exp = Math.toIntExact((long) json.get("exp"));
     }
 
     public JSONObject toJSON() {
